Cập nhật file lab2.2

+++ book.py
@@ -26,9 +26,17 @@ class Book:
 
     # DELETE
     def delete_book(self, db):
-        query = "DELETE FROM books WHERE book_id=%s"
-        db.execute_query(query, (self.book_id,))
-
+        checK_query = "DELETE FROM books WHERE book_id=%s"
+        row = db.execute_query(checK_query, (self.book_id,))
+        
+        if not row:
+            raise ValueError("Sách không tồn tại để xóa")
+        
+        if row[0] == 1:
+            raise ValueError("Không thể xóa. Sách này đang được mượn.")
+        
+        query = "DELETE FROM books WHERE book_id = %s"
+        db.execute_query(query,(self.book_id,))
     # READ / SEARCH
     @staticmethod
     def get_all_books(db):
